Move dataloaders to their own directory.
