+++ packages/orm/src/dataloaders/oneToManyDataLoader.ts
@@ -3,7 +3,7 @@ import Knex from "knex";
 import { Entity, getMetadata } from "../EntityManager";
 import { assertIdsAreTagged, deTagIds, getEm, maybeResolveReferenceToId, OneToManyCollection } from "../index";
 import { getOrSet, groupBy } from "../utils";
-import { LoaderCache } from "./EntityPersister";
+import { LoaderCache } from "../drivers/EntityPersister";
 
 export function oneToManyDataLoader<T extends Entity, U extends Entity>(
   knex: Knex,
